Added module overview page(s)
The links point to /module/[number], which doesn't exist yet, but should
be a page where you can configure the module.

The page loading time isn't exactly fast. This is, again, for the most
part caused by the lookup to matrix_sources. It might be an idea to cache
this view into a table or something...